Add files via upload
In this notebook, I have used a trained model to predict car prices. The trained model had been saved in lgbr_cars.model. I pulled the model from the file (see the notebook) and then predicted the price of a random car. Then I saved the model to Rest API (see server.py) and sent request to a server to make a prediction. As I used my notebook on the cloud, I wasn't able to use Rest API from localhost. Therefore, I saved The Model to serveo and sent request to serveo.